code generation for static-method-wrapper
This requires rudimentary implementations of both fn* and def.

Compared to tcljc's unit tests, it looks like the arity methods and
the def's member definition may swap places again.  For the moment,
the arity methods come first, then followed by the def member.  If it
stays like this, the swap will cause lots of noise in the unit tests.