refactor: move function code generation into a post processing step
A provisional change to swap def and method ordering in unit tests
back to that of tcljc.

In the future, all functions appearing in a namespace's top-level form
are subject to interdependence analysis followed by translation
pattern selection.  Prepare for this delayed code generation by only
collecting them for later in `new-function`, instead of generating
code on the spot.