refactor: always provide `consumed-type` to emit functions
The indicator "may insert return instruction" becomes the explicit
flag `then-leave?`.

The is necessary to pass the consumed type to different paths of an if
expression.  The paths cannot use their own idea of what is put on the
stack to generate the return instruction: one path may emit a RETURN
while the other emits an ARETURN.