refactor: do not special case polymorphic cast to void
Right now, the special case is for this unit test:

    (asm-expr [^java.lang.invoke.MethodHandle mh]
              ^void (.invoke mh))

This is in a auto-return-context, where the polymorphic return of the
invoke is modeled by the type void.  This makes the cast a "no
operation", and finally the functions's return type is taken to be the
logical type corresponding to polymorphic void: Object.  During code
generation, the type Object is then pushed down into the invoke.

There are two ways around this: change the function's return type to
void, or use (do ^void (.invoke mh) nil).  Both force the
non-polymorphic operand type void into the method invocation.