refactor: differentiate between computational and opaque constants
Anything that is of interest to constant folding is handled as
Constant: the non-void primitive types and Class.

Anything with a more complicated or non-unique internal representation
is mapped to Loadable: symbol, keyword, collection, or regex Pattern.

Move responsibility for loading into the objects, both as LDC or
constant pool entry.  Dito for conversion back to compile time data.