pull in some of the commented out tests
Small fixes and additions:

Add a def-ed JVM constant to the bindings, instead of an invoke of the
method that produces the constant's value.

Use more compact translation pattern for `when-not`, and support emit
of "branch on true", in addition to "branch on false".

Add value-type-exact method to Constant implementation.

When possible, use if_icmpne instead of lookupswitch in __arityOrNull.

Handle flag-statm-position for top-level expression, fn arity body,
and initial value of def.  Move try/catch/finally into closure when
not in flag-statm-position.