fix handling of polymorphic in "logical truth" context
For example, within an `if` expression the join type of polymorphic
and reference type T cannot be T, because this would mean the whole
`if` would be assigned type T.  This poses two problems: 1) The
polymorphic expression would be unconditionally cast to T, and 2) the
logical truth of the `if` expression would be miscompiled if T is
neither Object nor Boolean.

If in this scenario type T is primitive boolean, then a similar
situation arises.  The polymorphic expression must be evaluated as
Object, and then subjected to c.l.RT/booleanCast.