implement macros `and` and `or` as primitive functions
What makes this a bit hairy is that the arguments of these macros are
allowed to recur.  This means there are now two kinds of primitive
functions, the regular one whose arguments cannot recur, and the one
used for and/or.

Right now the translation pattern does not differ from that of the
original macro implementations.  But it may open up the path for more
compact translations, and it may serve as a template for a simpler
implementation of condgen's xcmp-n.