fix (compile|import)-nmsp not picking up new namespaces
Another case where a function closing over a mutable field causes it
to change by its own actions, and then not picking up the updated
value.

Maybe reject such a closure outright?  On the other hand, doing a lot
of work in a local function of a deftype's method is somewhat of an
anti-pattern.  It is a sharp edge anyway and should not be encouraged.