ensure visibility of ephemeral classes across namespace segments
While compiling a namespace, an ephemeral class exists from the point
of definition up to the end of the namespace.  Its scope of visibilty
can span multiple namespace segments, or it ends early because it
appears in top-level forms that are discarded before being recompiled
again.  The class instance appears in method signatures, namespace
defs, class definitions, and the imports map.

Ephemeral class definitions appearing later in the namespace may refer
to ephemerals defined earlier.  Even if the earlier definitions have
already been finalized and written to the dynamic class loader, the
new ephemeral must link to the original ephemerals and not application
class instances.  Otherwise, type checks may fail.

The example in question is a `reify*` based on a class from a prior
namespace segment.  The ephemeral class loader must resolve the base
class to the prior segment's ephemeral class, *not* the one from the
application's class loader.

The solution is to stack ephemeral class loaders.  A segment's
ephemeral loader uses the prior ephemeral loader as parent.  This way,
classes can be redefined in case of a rollback within a segment, but
otherwise resolve to an earlier segment's ephemeral.