change `if` to take exactly three arguments
If a polymorphic empty path is desired, the empty expression `(do)`
should be used.

There is a slight wrinkle to this: `(do)` gets assigned the same type
as a polymorphic method invocation.  Its join-type is at least Object,
to prevent locical-truth from pushing an arbitrary reference result
type into the invocation.  This contradicts the expected behaviour of
`if`, that the empty path does not contribute to the if expression's
overall type.

The workaround for now is to special case `emitter-empty` paths for if
expressions.