refactor: tweak `invoke-value` to always use Object as return type
The expression signals this type, and only Object is ever used as the
dynamic invocation's return type.

Most of the type adaption is then taken over by the regular emit
functions, with the exception of consumed type `boolean`.  As before,
a call to `booleanCast` is appended in this case to implement
Clojure's idea of a logically true value.