push adaption of reference type into paths
This may lead to a certain amount of code duplication, when a single
shared adaption after the if/switch/try is turned into several
distinct adaption operations along each path.

In practice, the effect seems to be negligible: using
tcljx-stage2.mdir-xpl for comparison, adding the test increases the
byte count by 17, with the predicate's bytecode being part of this
number.

There is also an upside: A failed type adaption, usually a checkcast,
now happens in the context of a specific path, showing a more precise
line number in the stack trace.